+++ admin/words.js
@@ -2,40 +2,148 @@
 "use strict";
 
 systemDictionary = {
-	"synochat adapter settings": {
-		"en": "Adapter settings for synochat",
-		"de": "Adaptereinstellungen für synochat",
-		"ru": "Настройки адаптера для synochat",
-		"pt": "Configurações do adaptador para synochat",
-		"nl": "Adapterinstellingen voor synochat",
-		"fr": "Paramètres d'adaptateur pour synochat",
-		"it": "Impostazioni dell'adattatore per synochat",
-		"es": "Ajustes del adaptador para synochat",
-		"pl": "Ustawienia adaptera dla synochat",
-		"zh-cn": "synochat的适配器设置"
+	"Main settings": {
+		"en": "Main settings",
+		"de": "Haupteinstellungen",
+		"ru": "Основные параметры",
+		"pt": "Configurações principais",
+		"nl": "Belangrijkste instellingen",
+		"fr": "Réglages principaux",
+		"it": "Impostazioni principali",
+		"es": "Ajustes principales",
+		"pl": "Ustawienia główne",
+		"zh-cn": "主要设置"
 	},
-	"option1": {
-		"en": "option1",
-		"de": "Option 1",
-		"ru": "Опция 1",
-		"pt": "Opção 1",
-		"nl": "Optie 1",
-		"fr": "Option 1",
-		"it": "opzione 1",
-		"es": "Opción 1",
-		"pl": "opcja 1",
-		"zh-cn": "选项1"
-	},
-	"option2": {
-		"en": "option2",
-		"de": "Option 2",
-		"ru": "вариант 2",
-		"pt": "opção 2",
-		"nl": "Optie 2",
-		"fr": "Option 2",
-		"it": "opzione 2",
-		"es": "opcion 2",
-		"pl": "Opcja 2",
-		"zh-cn": "选项2"
-	}
+	"Channel management": {
+		"en": "Channel management",
+		"de": "Kanalverwaltung",
+		"ru": "Управление каналами",
+		"pt": "Gestão de canais",
+		"nl": "Kanaal beheer",
+		"fr": "La gestion des canaux",
+		"it": "Gestione del canale",
+		"es": "Gestión de canales",
+		"pl": "Zarządzanie kanałami",
+		"zh-cn": "渠道管理"
+	  },
+	  "Help": {
+		"en": "Help",
+		"de": "Hilfe",
+		"ru": "Помощь",
+		"pt": "Ajuda",
+		"nl": "Helpen",
+		"fr": "Aider",
+		"it": "Aiuto",
+		"es": "Ayudar",
+		"pl": "Pomoc",
+		"zh-cn": "帮助"
+	  },
+	  "Synology URL/IP": {
+		"en": "Synology URL/IP",
+		"de": "Synology-URL/IP",
+		"ru": "URL-адрес/IP-адрес Synology",
+		"pt": "URL/IP da Synology",
+		"nl": "Synology URL/IP",
+		"fr": "URL/IP de Synology",
+		"it": "Synology URL/IP",
+		"es": "URL/IP de Synology",
+		"pl": "Adres URL/IP firmy Synology",
+		"zh-cn": "Synology URL/IP"
+	  },
+	  "Validate SSL certificate": {
+		"en": "Validate SSL certificate",
+		"de": "SSL-Zertifikat validieren",
+		"ru": "Проверить SSL-сертификат",
+		"pt": "Validar certificado SSL",
+		"nl": "SSL-certificaat valideren",
+		"fr": "Valider le certificat SSL",
+		"it": "Convalida il certificato SSL",
+		"es": "Validar certificado SSL",
+		"pl": "Zweryfikuj certyfikat SSL",
+		"zh-cn": "验证 SSL 证书"
+	  },
+	  "Hostname / IP address of ioBroker instance": {
+		"en": "Hostname / IP address of ioBroker instance",
+		"de": "Hostname / IP-Adresse der ioBroker-Instanz",
+		"ru": "Имя хоста/IP-адрес экземпляра ioBroker",
+		"pt": "Nome do host/endereço IP da instância ioBroker",
+		"nl": "Hostnaam / IP-adres van ioBroker-instantie",
+		"fr": "Nom d'hôte / adresse IP de l'instance ioBroker",
+		"it": "Nome host/indirizzo IP dell'istanza ioBroker",
+		"es": "Nombre de host/dirección IP de la instancia de ioBroker",
+		"pl": "Nazwa hosta / adres IP instancji ioBroker",
+		"zh-cn": "ioBroker 实例的主机名/IP 地址"
+	  },
+	  "Web instance for messages send from Synology chat to ioBroker adapter instance": {
+		"en": "Web instance for messages send from Synology chat to ioBroker adapter instance",
+		"de": "Webinstanz für Nachrichten, die vom Synology-Chat an die ioBroker-Adapterinstanz gesendet werden",
+		"ru": "Веб-экземпляр для отправки сообщений из чата Synology в экземпляр адаптера ioBroker",
+		"pt": "Instância da Web para mensagens enviadas do chat Synology para a instância do adaptador ioBroker",
+		"nl": "Webinstantie voor berichten die worden verzonden van Synology-chat naar instantie van ioBroker-adapter",
+		"fr": "Instance Web pour les messages envoyés depuis le chat Synology vers l'instance de l'adaptateur ioBroker",
+		"it": "Istanza Web per i messaggi inviati dalla chat di Synology all'istanza dell'adattatore ioBroker",
+		"es": "Instancia web para mensajes enviados desde el chat de Synology a la instancia del adaptador ioBroker",
+		"pl": "Instancja internetowa dla wiadomości wysyłanych z czatu Synology do instancji adaptera ioBroker",
+		"zh-cn": "从 Synology 聊天发送到 ioBroker 适配器实例的消息的 Web 实例"
+	  },
+	  "NOTE: The WebHook URL for receiving data from Synology chat server can be found as object value of 'ADAPTER_INSTANCE.info.webHookUrl' for this adapter instance.": {
+		"en": "NOTE:\nThe WebHook URL for receiving data from Synology chat server can be found as object value of 'ADAPTER_INSTANCE.info.webHookUrl' for this adapter instance.",
+		"de": "HINWEIS:\nDie WebHook-URL zum Empfangen von Daten vom Synology-Chatserver kann als Objektwert von „ADAPTER_INSTANCE.info.webHookUrl“ für diese Adapterinstanz gefunden werden.",
+		"ru": "ПРИМЕЧАНИЕ:\nURL-адрес WebHook для получения данных от сервера чата Synology можно найти как значение объекта ADAPTER_INSTANCE.info.webHookUrl для этого экземпляра адаптера.",
+		"pt": "NOTA:\nA URL do WebHook para receber dados do servidor de bate-papo Synology pode ser localizada como valor de objeto de 'ADAPTER_INSTANCE.info.webHookUrl' para esta instância do adaptador.",
+		"nl": "NOTITIE:\nDe WebHook-URL voor het ontvangen van gegevens van de Synology-chatserver kan worden gevonden als objectwaarde van 'ADAPTER_INSTANCE.info.webHookUrl' voor deze adapterinstantie.",
+		"fr": "REMARQUE:\nL'URL WebHook pour recevoir des données du serveur de chat Synology peut être trouvée en tant que valeur d'objet de 'ADAPTER_INSTANCE.info.webHookUrl' pour cette instance d'adaptateur.",
+		"it": "NOTA:\nL'URL WebHook per la ricezione dei dati dal server di chat Synology può essere trovato come valore oggetto di 'ADAPTER_INSTANCE.info.webHookUrl' per questa istanza dell'adattatore.",
+		"es": "NOTA:\nLa URL de WebHook para recibir datos del servidor de chat de Synology se puede encontrar como valor de objeto de 'ADAPTER_INSTANCE.info.webHookUrl' para esta instancia de adaptador.",
+		"pl": "UWAGA:\nAdres URL elementu WebHook do odbierania danych z serwera czatu firmy Synology można znaleźć jako wartość obiektu „ADAPTER_INSTANCE.info.webHookUrl” dla tej instancji adaptera.",
+		"zh-cn": "笔记：\n用于从 Synology 聊天服务器接收数据的 WebHook URL 可以作为此适配器实例的“ADAPTER_INSTANCE.info.webHookUrl”的对象值找到。"
+	  },
+	  "Channel name": {
+		"en": "Channel name",
+		"de": "Kanal Name",
+		"ru": "Название канала",
+		"pt": "Nome do canal",
+		"nl": "Kanaal naam",
+		"fr": "Nom du canal",
+		"it": "Nome del canale",
+		"es": "Nombre del Canal",
+		"pl": "Nazwa kanału",
+		"zh-cn": "频道名称"
+	  },
+	  "Access token": {
+		"en": "Access token",
+		"de": "Zugangstoken",
+		"ru": "Токен доступа",
+		"pt": "Token de acesso",
+		"nl": "Toegangstoken",
+		"fr": "Jeton d'accès",
+		"it": "Token di accesso",
+		"es": "ficha de acceso",
+		"pl": "Token dostępu",
+		"zh-cn": "访问令牌"
+	  },
+	  "Channel type": {
+		"en": "Channel type",
+		"de": "Kanaltyp",
+		"ru": "Тип канала",
+		"pt": "Tipo de canal",
+		"nl": "Kanaaltype",
+		"fr": "Type de canal",
+		"it": "Tipo di canale",
+		"es": "Tipo de canal",
+		"pl": "Typ kanału",
+		"zh-cn": "渠道类型"
+	  },
+	  "Channel enabled": {
+		"en": "Channel enabled",
+		"de": "Kanal aktiviert",
+		"ru": "Канал включен",
+		"pt": "Canal ativado",
+		"nl": "Kanaal ingeschakeld",
+		"fr": "Canal activé",
+		"it": "Canale abilitato",
+		"es": "Canal habilitado",
+		"pl": "Kanał włączony",
+		"zh-cn": "启用频道"
+	  }
 };